# inServ/src/inserv/models/entity_ref.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from inserv.models.base_model import Model
from inserv import util


class EntityRef(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, id=None, href=None, name=None):  # noqa: E501
        """EntityRef - a model defined in OpenAPI

        :param id: The id of this EntityRef.  # noqa: E501
        :type id: str
        :param href: The href of this EntityRef.  # noqa: E501
        :type href: str
        :param name: The name of this EntityRef.  # noqa: E501
        :type name: str
        """
        self.openapi_types = {
            'id': str,
            'href': str,
            'name': str
        }

        self.attribute_map = {
            'id': 'id',
            'href': 'href',
            'name': 'name'
        }

        self._id = id
        self._href = href
        self._name = name

    @classmethod
    def from_dict(cls, dikt) -> 'EntityRef':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The EntityRef of this EntityRef.  # noqa: E501
        :rtype: EntityRef
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """Gets the id of this EntityRef.


        :return: The id of this EntityRef.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this EntityRef.


        :param id: The id of this EntityRef.
        :type id: str
        """

        self._id = id

    @property
    def href(self) -> str:
        """Gets the href of this EntityRef.


        :return: The href of this EntityRef.
        :rtype: str
        """
        return self._href

    @href.setter
    def href(self, href: str):
        """Sets the href of this EntityRef.


        :param href: The href of this EntityRef.
        :type href: str
        """

        self._href = href

    @property
    def name(self) -> str:
        """Gets the name of this EntityRef.


        :return: The name of this EntityRef.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this EntityRef.


        :param name: The name of this EntityRef.
        :type name: str
        """

        self._name = name