# inServ/src/inserv/models/error.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from inserv.models.base_model import Model
from inserv import util


class Error(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, code=None, reason=None):  # noqa: E501
        """Error - a model defined in OpenAPI

        :param code: The code of this Error.  # noqa: E501
        :type code: str
        :param reason: The reason of this Error.  # noqa: E501
        :type reason: str
        """
        self.openapi_types = {
            'code': str,
            'reason': str
        }

        self.attribute_map = {
            'code': 'code',
            'reason': 'reason'
        }

        self._code = code
        self._reason = reason

    @classmethod
    def from_dict(cls, dikt) -> 'Error':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Error of this Error.  # noqa: E501
        :rtype: Error
        """
        return util.deserialize_model(dikt, cls)

    @property
    def code(self) -> str:
        """Gets the code of this Error.


        :return: The code of this Error.
        :rtype: str
        """
        return self._code

    @code.setter
    def code(self, code: str):
        """Sets the code of this Error.


        :param code: The code of this Error.
        :type code: str
        """

        self._code = code

    @property
    def reason(self) -> str:
        """Gets the reason of this Error.


        :return: The reason of this Error.
        :rtype: str
        """
        return self._reason

    @reason.setter
    def reason(self, reason: str):
        """Sets the reason of this Error.


        :param reason: The reason of this Error.
        :type reason: str
        """

        self._reason = reason