# inServ/src/inserv/models/time_period.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from inserv.models.base_model import Model
from inserv import util


class TimePeriod(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, start_date_time=None, end_date_time=None):  # noqa: E501
        """TimePeriod - a model defined in OpenAPI

        :param start_date_time: The start_date_time of this TimePeriod.  # noqa: E501
        :type start_date_time: datetime
        :param end_date_time: The end_date_time of this TimePeriod.  # noqa: E501
        :type end_date_time: datetime
        """
        self.openapi_types = {
            'start_date_time': datetime,
            'end_date_time': datetime
        }

        self.attribute_map = {
            'start_date_time': 'startDateTime',
            'end_date_time': 'endDateTime'
        }

        self._start_date_time = start_date_time
        self._end_date_time = end_date_time

    @classmethod
    def from_dict(cls, dikt) -> 'TimePeriod':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The TimePeriod of this TimePeriod.  # noqa: E501
        :rtype: TimePeriod
        """
        return util.deserialize_model(dikt, cls)

    @property
    def start_date_time(self) -> datetime:
        """Gets the start_date_time of this TimePeriod.


        :return: The start_date_time of this TimePeriod.
        :rtype: datetime
        """
        return self._start_date_time

    @start_date_time.setter
    def start_date_time(self, start_date_time: datetime):
        """Sets the start_date_time of this TimePeriod.


        :param start_date_time: The start_date_time of this TimePeriod.
        :type start_date_time: datetime
        """

        self._start_date_time = start_date_time

    @property
    def end_date_time(self) -> datetime:
        """Gets the end_date_time of this TimePeriod.


        :return: The end_date_time of this TimePeriod.
        :rtype: datetime
        """
        return self._end_date_time

    @end_date_time.setter
    def end_date_time(self, end_date_time: datetime):
        """Sets the end_date_time of this TimePeriod.


        :param end_date_time: The end_date_time of this TimePeriod.
        :type end_date_time: datetime
        """

        self._end_date_time = end_date_time